Change png to react icons on the video-conference page
