+++ package.json
@@ -24,6 +24,7 @@
     "react-calendar": "^5.0.0",
     "react-dom": "^18",
     "react-hook-form": "^7.52.0",
+    "react-icons": "^5.3.0",
     "react-intl": "^6.6.8",
     "tailwind-merge": "^2.5.2"
   },
